one more configuration removed

+++ train/intents_training.py
@@ -18,7 +18,44 @@ class IntentsTraining:
         self.__recorder_recognizer = RecorderRecognizer()
 
     def train_responses(self, assistant, intent_type):
-        recognized_data = ''
+        assistant.say_current_text('Начинаем разговор. Я вас слушаю')
+        assistant.say_current_text('Скажите стоп, чтобы закончить тренировку')
+
+        while (True):
+            listened_audio = self.__recorder_recognizer.record_audio()
+            recognized_data = self.__recorder_recognizer.recognize_audio(listened_audio)
+
+            if recognized_data == 'Стоп':
+                assistant.say_current_text('Отлично! Будет скучно, возвращайся')
+                break
+
+            if path.exists('../json/intents.json'):
+                with open('../json/intents.json', encoding='utf-8') as file:
+                    json_data = load(file)
+
+            elif path.exists('json/intents.json'):
+                with open('json/intents.json', encoding='utf-8') as file:
+                    json_data = load(file)
+
+                if recognized_data not in json_data[intent_type]["response"] and recognized_data:
+                    json_data[intent_type]["response"].append(recognized_data)
+                    print(recognized_data + " added to 'intents.json' [responses]")
+
+                else:
+                    continue
+
+            if path.exists('../json/intents.json'):
+                with open('../json/intents.json', mode='w', encoding='utf-8') as file:
+                    dump(json_data, file, ensure_ascii=False, indent=2)
+
+            elif path.exists('json/intents.json'):
+                with open('json/intents.json', mode='w', encoding='utf-8') as file:
+                    dump(json_data, file, ensure_ascii=False, indent=2)
+
+            else:
+                break
+
+    def train_answers(self, assistant, intent_type):
         assistant.say_current_text('Начинаем разговор. Я вас слшуаю')
 
         while (True):
@@ -32,32 +69,8 @@ class IntentsTraining:
                 with open('../json/intents.json', encoding='utf-8') as file:
                     json_data = load(file)
 
-                if recognized_data not in json_data[intent_type]["response"] and recognized_data:
-                    json_data[intent_type]["response"].append(recognized_data)
-                    print(recognized_data + " added to 'intents.json' [responses]")
-
-                else:
-                    continue
-
-                with open('../json/intents.json', mode='w', encoding='utf-8') as file:
-                    dump(json_data, file, ensure_ascii=False, indent=2)
-
-            else:
-                break
-
-    def train_answers(self, assistant, intent_type):
-        recognized_data = ''
-        assistant.say_current_text('Начинаем разговор. Я вас слшуаю')
-
-        while (True):
-            listened_audio = self.__recorder_recognizer.record_audio()
-            recognized_data = self.__recorder_recognizer.recognize_audio(listened_audio)
-
-            if recognized_data == 'Стоп':
-                break
-
-            if path.exists('../json/intents.json'):
-                with open('../json/intents.json', encoding='utf-8') as file:
+            elif path.exists('json/intents.json'):
+                with open('json/intents.json', encoding='utf-8') as file:
                     json_data = load(file)
 
 
@@ -68,7 +81,12 @@ class IntentsTraining:
                 else:
                     continue
 
+            if path.exists('../json/intents.json'):
                 with open('../json/intents.json', mode='w', encoding='utf-8') as file:
+                    dump(json_data, file, ensure_ascii=False, indent=2)
+
+            elif path.exists('json/intents.json'):
+                with open('json/intents.json', mode='w', encoding='utf-8') as file:
                     dump(json_data, file, ensure_ascii=False, indent=2)
 
             else:
@@ -77,7 +95,7 @@ class IntentsTraining:
     @staticmethod
     def now_available_intents(assistant):
         assistant.say_current_text('На данный момент в моём разуме приисутсвует понимание следующих вещей:')
-        with open('../json/intents.json', encoding='utf-8') as file:
+        with open('json/intents.json', encoding='utf-8') as file:
             json_data = load(file)
 
         for index, key in enumerate(json_data.keys(), 1):
